Fixed race condition
There was a race condition where it would sometimes delete and create
cache file, whilst another instance deletes the new cache file instead
of the old one, due to it using the same name. As a result the first
instance would try to fetch the newly created cache file, only to find
it's been deleted by the second instance and can't display the image.

Fixed by creating a lock on the index.html file within the cache
directory, and if lock is successfuly only then delete the expired
cache files, therefore this only happens once at a time no matter how
many instances are runing.

+++ thumb.php
@@ -72,17 +72,22 @@ $file_name = THUMB_CACHE . $file_hash . '.img.txt';
 if(!file_exists(THUMB_CACHE . 'index.html')) {
     touch(THUMB_CACHE . 'index.html');
 }
-if(time()-THUMB_CACHE_AGE>filemtime(THUMB_CACHE . 'index.html')) {
-    $files = glob(THUMB_CACHE . '*.img.txt');
-    if(is_array($files)&&count($files)>0) {
-        foreach($files as $file) {
-            if(time()-THUMB_CACHE_AGE>filemtime($file)) {
-                unlink($file);
+$fp = fopen(THUMB_CACHE . 'index.html','r');
+if(flock($fp,LOCK_EX)) {
+    if(time()-THUMB_CACHE_AGE>filemtime(THUMB_CACHE . 'index.html')) {
+        $files = glob(THUMB_CACHE . '*.img.txt');
+        if(is_array($files)&&count($files)>0) {
+            foreach($files as $file) {
+                if(time()-THUMB_CACHE_AGE>filemtime($file)) {
+                    unlink($file);
+                }
             }
         }
+        touch(THUMB_CACHE . 'index.html');
     }
-    touch(THUMB_CACHE . 'index.html');
+    flock($fp,LOCK_UN);
 }
+fclose($fp);
 
 if(THUMB_BROWSER_CACHE&&(isset($_SERVER['HTTP_IF_MODIFIED_SINCE'])||isset($_SERVER['HTTP_IF_NONE_MATCH']))) {
     if($_SERVER['HTTP_IF_MODIFIED_SINCE']==$file_date&&$_SERVER['HTTP_IF_NONE_MATCH']==$file_hash) {
